# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 606a4e64a3e0491123244fdd
  title: SchoolManagementSystem
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /exam:
    post:
      x-skaffolder-id: 606a5279a3e04911232450b4
      x-skaffolder-name: create
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - exam
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                place:
                  type: string
                score:
                  type: number
                valid:
                  type: boolean
    get:
      x-skaffolder-id: 606a5279a3e04911232450ba
      x-skaffolder-name: list
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - exam
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/{id}":
    delete:
      x-skaffolder-id: 606a5279a3e04911232450bb
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 606a5279a3e04911232450b6
      x-skaffolder-name: get
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
    post:
      x-skaffolder-id: 606a5279a3e04911232450b5
      x-skaffolder-name: update
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/exam"
  "/exam/findBy_course/{key}":
    get:
      x-skaffolder-id: 606a5271a3e04911232450aa
      x-skaffolder-name: findBy_course
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_course
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_student/{key}":
    get:
      x-skaffolder-id: 606a5261a3e0491123245091
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 606a526aa3e049112324509e
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/validate/{id}":
    post:
      x-skaffolder-id: 606a5342a3e04911232450bd
      x-skaffolder-name: validate
      x-skaffolder-id-resource: 606a5147a3e0491123245082
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: 
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Boolean
      x-skaffolder-url: /validate/{id}
      x-skaffolder-roles:
      summary: 
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: String
          in: path
          description: id of the exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /students:
    post:
      x-skaffolder-id: 606a5261a3e049112324508c
      x-skaffolder-name: create
      x-skaffolder-id-resource: 606a5044a3e049112324506e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dateofbirth:
                  type: integer
                name:
                  type: string
                surname:
                  type: string
    get:
      x-skaffolder-id: 606a5261a3e0491123245095
      x-skaffolder-name: list
      x-skaffolder-id-resource: 606a5044a3e049112324506e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
  "/students/{id}":
    delete:
      x-skaffolder-id: 606a5261a3e0491123245096
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 606a5044a3e049112324506e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 606a5261a3e049112324508e
      x-skaffolder-name: get
      x-skaffolder-id-resource: 606a5044a3e049112324506e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
    post:
      x-skaffolder-id: 606a5261a3e049112324508d
      x-skaffolder-name: update
      x-skaffolder-id-resource: 606a5044a3e049112324506e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/student"
  "/students/findBy_courses/{key}":
    get:
      x-skaffolder-id: 606a5272a3e04911232450ae
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 606a5044a3e049112324506e
      x-skaffolder-resource: student
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - student
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    exam:
      x-skaffolder-id: 606a5147a3e0491123245082
      x-skaffolder-id-db: 606a4e64a3e0491123244fde
      x-skaffolder-id-entity: 606a5146a3e049112324507f
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        place:
          type: string
          x-skaffolder-id-attr: 606a5146a3e0491123245081
          x-skaffolder-type: String
        score:
          type: number
          x-skaffolder-id-attr: 606a5146a3e0491123245080
          x-skaffolder-type: Number
        valid:
          type: boolean
          x-skaffolder-id-attr: 606a53e8a3e04911232450bf
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _course:
          x-skaffolder-id: 606a51a7a3e0491123245086
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 606a5146a3e049112324507f
          x-skaffolder-ent2: 606a511da3e049112324507a
        _student:
          x-skaffolder-id: 606a51a7a3e0491123245084
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 606a5146a3e049112324507f
          x-skaffolder-ent2: 606a5044a3e049112324506d
        _teacher:
          x-skaffolder-id: 606a51a7a3e0491123245085
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 606a5146a3e049112324507f
          x-skaffolder-ent2: 606a50f3a3e0491123245074
      required:
        - _id
    student:
      x-skaffolder-id: 606a5044a3e049112324506e
      x-skaffolder-id-db: 606a4e64a3e0491123244fde
      x-skaffolder-id-entity: 606a5044a3e049112324506d
      x-skaffolder-url: /students
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        dateofbirth:
          type: integer
          x-skaffolder-id-attr: 606a50aea3e0491123245072
          x-skaffolder-type: Date
        name:
          type: string
          x-skaffolder-id-attr: 606a50aea3e0491123245070
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 606a50aea3e0491123245071
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 606a5213a3e0491123245088
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 606a5044a3e049112324506d
          x-skaffolder-ent2: 606a511da3e049112324507a
      required:
        - _id
...